Allow to check next part and populate MIME content even error
Current email parser library has the ability to ignore MIME
with bad ending boundary. To avoid this function being erased
or interrupt empty/corrupt MIME header detection, a new func
is added to distinguish whether the next part is just EOF or
it is an empty/corrupt MIME header.

Most golang stdlib will return nil object when error occur. This
causes problem certain can be ignored but because object is nil,
ignoring the error seem not helpful. This change will allow the
MIME content to be returned too even when MIME header not found -
which can cause error.