Use modified pkg net/textproto
This modified pkg net/textproto will still parse any readable MIME
header and also return the same error to user. This allow user
to choose whether to ignore the error or not. Previous library
(stdlib) will stop parsing when encounter any error. Any good MIME
that follow behind can not be parsed. With this modification, as
long as the MIME has good header then it will be parsed.